+++ feed/src/main/java/com/muvi/feed/di/FeedComponent.kt
@@ -2,16 +2,12 @@ package com.muvi.feed.di
 
 import androidx.lifecycle.ViewModelProviders
 import com.muvi.core.di.CoreComponent
-import com.muvi.feed.BuildConfig
 import com.muvi.feed.FeedFragment
 import com.muvi.feed.FeedViewModel
 import com.muvi.feed.FeedViewModelFactory
 import com.muvi.feed_cache.di.FeedCacheModule
 import com.muvi.feed_data.di.FeedDataModule
-import com.muvi.feed_remote.di.DebugMode
 import com.muvi.feed_remote.di.FeedRemoteModule
-import com.muvi.feed_remote.di.LetterboxdApiKey
-import com.muvi.feed_remote.di.LetterboxdApiSecret
 import dagger.BindsInstance
 import dagger.Component
 import dagger.Module
@@ -49,18 +45,4 @@ internal object FeedModule {
     fun viewModel(feedFragment: FeedFragment, viewModelFactory: FeedViewModelFactory) =
             ViewModelProviders.of(feedFragment, viewModelFactory).get(FeedViewModel::class.java)
 
-    @JvmStatic
-    @LetterboxdApiKey
-    @Provides
-    fun letterboxdApiKey() = BuildConfig.API_KEY
-
-    @JvmStatic
-    @LetterboxdApiSecret
-    @Provides
-    fun letterboxdApiSecret() = BuildConfig.API_SECRET
-
-    @JvmStatic
-    @DebugMode
-    @Provides
-    fun debugMode() = BuildConfig.DEBUG
 }
