+++ film_detail/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     implementation project(':film_detail_data')
     implementation project(':film_detail_domain')
     implementation project(':film_detail_remote')
+    compileOnly project(':letterboxd_api')
     implementation project(':navigation')
 
     implementation libraries.androidxAppCompat
